backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0,<3.16.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
psycopg2-binary>=2.9.9,<3.0.0
python-dotenv>=1.0.1,<2.0.0
pytest>=8.0.0,<9.0.0
pytest-django>=4.8.0,<5.0.0
factory-boy>=3.3.0,<4.0.0
django-ratelimit>=4.1.0,<5.0.0
drf-spectacular>=0.27.0,<0.29.0
structlog>=24.1.0,<25.0.0
django-redis>=5.4.0,<6.0.0
